Update clang-tidy format
